added swatmultimodel class and improved documentation

+++ pySWATPlus/PymooBestSolution.py
@@ -15,17 +15,17 @@ this.lock = multiprocessing.Lock()
 
 def add_solutions(paths_array: List[Dict[str, str]], errors_array: np.ndarray) -> None:
     """
-    Update the best solution based on the provided paths and errors.
+    Update the best solution based on the provided paths and errors. Only the best solution is kept, the other ones are deleted.
 
     Parameters:
-    - paths_array (np.ndarray): An array of paths.
+    - paths_array (np.ndarray): An array of dictionaries containing the paths to the output files.
     - errors_array (np.ndarray): An array of errors.
 
     Returns:
     None
     """
     
-    min_idx = np.argmin(errors_array)
+    min_idx = np.nanargmin(errors_array)
     path = paths_array[min_idx]
     error = errors_array[min_idx]
 
@@ -36,7 +36,7 @@ def add_solutions(paths_array: List[Dict[str, str]], errors_array: np.ndarray) -
 
     
     for i in all_paths:
-        if i not in best:
+        if i not in best and i is not None:
             shutil.rmtree(i)
 
 
@@ -53,10 +53,12 @@ def add_solution(X: np.ndarray, path: Dict[str, str], error: float) -> None:
     None
     """
     with this.lock:
+
         if this.error is None or this.error > error:
             this.X = X
             this.path = path
             this.error = error
+
 
 
 def get_solution() -> Tuple[np.ndarray, Dict[str, str], float]:
@@ -64,7 +66,7 @@ def get_solution() -> Tuple[np.ndarray, Dict[str, str], float]:
     Retrieve the best solution.
 
     Returns:
-    Tuple[Optional[Any], Optional[str], Optional[float]]: The best solution, path, and error.
+    Tuple[np.ndarray, Dict[str, str], float]: The best solution, path, and error.
     """
     
     with this.lock:
